pyincept-37: Refactoring: incremental progress towards a common superclass for testing all archetype output.
